complete checkpoint 2 (kinda)
I was kind of confused on checkpoint 2. I was expecting
the console.log to print out the same results as we got
in exercise 1 but instead it printed '[Function: incrementor]'
10 times onto the console.
Was this the expected result? Or did I just misread the directions?
Any help would be greatly appreciated on this.
Thanks,
Alex